Schema validation for change files
