+++ packages/chronus/src/config/parse.ts
@@ -1,7 +1,6 @@
-import z, { ZodError, type ZodIssue } from "zod";
+import z from "zod";
 import type { TextFile } from "../file/types.js";
-import { ChronusDiagnosticError, type Diagnostic } from "../utils/errors.js";
-import { getLocationInYamlScript, parseYaml, type YamlFile } from "../yaml/index.js";
+import { parseYaml, validateYamlFile } from "../yaml/index.js";
 import type { ChronusUserConfig } from "./types.js";
 
 const versionPolicySchema = z.union([
@@ -36,21 +35,5 @@ const schema = z.object({
 
 export function parseConfig(content: string | TextFile): ChronusUserConfig {
   const parsed = parseYaml(content);
-  try {
-    return schema.parse(parsed.data);
-  } catch (e) {
-    if (e instanceof ZodError) {
-      throw new ChronusDiagnosticError(e.errors.map((issue) => convertZodIssueToDiagnostic(issue, parsed)));
-    }
-    throw e;
-  }
-}
-
-function convertZodIssueToDiagnostic(issue: ZodIssue, file: YamlFile): Diagnostic {
-  return {
-    code: `yaml-${issue.code.toLowerCase().replace(/_/g, "-")}`,
-    message: issue.message,
-    severity: "error",
-    target: getLocationInYamlScript(file, issue.path),
-  };
+  return validateYamlFile(parsed, schema);
 }
