+++ packages/chronus/src/yaml/index.ts
@@ -1,3 +1,4 @@
 export { getLocationInYamlScript } from "./location.js";
 export { parseYaml } from "./parse.js";
+export { validateYamlFile } from "./schema-validator.js";
 export type { YamlFile } from "./types.js";
